# api/openapi.yaml
openapi: 3.1.0
info:
  title: Catalift API
  version: 1.7.0
  description: |
    The REST API the Catalift web app uses: sellers upload products and start generation, reviewers edit,
    regenerate, approve and export listings, and both see AI cost and the budget state (ADR-0007).
    Serves: US-00-001 to US-00-012. Style: docs/api/README.md. The server default for X-API-Version is the
    newest date version, currently 2026-10-04.
  contact:
    name: Catalift team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-api-style: >-
    REST over HTTPS with JSON bodies, plus multipart for CSV and image uploads and text/csv for export
    downloads; one Go service exposing five areas (auth, catalogue, listings, export, cost) under one /v1.
  x-versioning: >-
    Major in the path (/v1, ADR-0007); dated behaviour changes inside a major in X-API-Version, server
    default the newest (2026-10-04).
  x-authentication: >-
    A server-side session in an HTTP-only secure cookie (catalift_session) set by POST /v1/sessions, with
    an X-CSRF-Token header on every state-changing request (ADR-0006). Roles seller and reviewer are
    checked on every route; a seller calling a reviewer action gets 403 forbidden_role.
  x-conventions:
    - rule: 'The base path is /v1, not /api/v1.'
      why: 'ADR-0007 and HLD section 5 fixed /v1; the reverse proxy forwards /v1 and serves the web app at /.'
...
    - rule: 'Ids are decimal strings of the database''s bigint ids.'
...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /listings:
    get:
      operationId: listListings
      summary: The review grid
      description: One row per product and channel (AC-US-00-007-1), by product SKU then channel.
      tags: [listings]
      x-story-ids: [US-00-004, US-00-007, US-00-009]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/UploadFilter'
        - name: filter[channel]
          in: query
          description: Only this channel (AC-US-00-007-5; sequential scan by design, under 10^4 rows).
          schema: { type: string, pattern: '^[a-z][a-z0-9_]{1,39}$' }
        - name: filter[rule_status]
          in: query
          description: Only this rule status (AC-US-00-007-5; sequential scan by design).
          schema: { $ref: '#/components/schemas/RuleStatus' }
        - name: filter[approved]
          in: query
          description: Only approved (true) or unapproved (false) listings.
          schema: { type: boolean }
      responses:
        '200':
          description: One page of grid rows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /listings/{listing_id}:
    parameters:
      - $ref: '#/components/parameters/ListingId'
    get:
      operationId: getListing
      summary: Get one listing with its rule failures
      tags: [listings]
      x-story-ids: [US-00-004, US-00-007]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The listing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Listing' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      operationId: updateListing
      summary: Edit listing text (reviewer)
      description: >-
        Saves the given fields, raises the version, clears any approval and re-checks the rules in one transaction
        (AC-US-00-007-2 and -4, AC-US-00-004-6). The body carries the version it was made against; a stale
        version is 409 version_conflict, so a retry cannot apply twice.
      tags: [listings]
      x-story-ids: [US-00-007, US-00-004]
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ListingUpdate' }
            example: { version: 2, title: 'Women''s Black Cotton Kurta, Regular Fit' }
      responses:
        '200':
          description: The listing after the edit and re-check
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Listing' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /approvals:
    post:
      operationId: createApprovals
      summary: Approve many listings (reviewer)
      description: >-
        Approves each listed (listing_id, version) whose listing is still at that version and passing its rules, in
        one statement per listing; the rest are skipped with a reason (AC-US-00-009-1 to -3, Q-009).
      tags: [listings]
      x-story-ids: [US-00-009]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalBatchCreate' }
            example:
              items:
                - { listing_id: '412', version: 3 }
                - { listing_id: '413', version: 1 }
      responses:
        '201':
          description: Approved and skipped listings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalBatchResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /rule-checks:
    post:
      operationId: createRuleCheck
      summary: Check draft listing text against its channel's rules (reviewer)
      description: >-
        Runs the rules engine on a listing with the given fields replaced, and saves nothing, so the editor's badge
        updates as the reviewer types (v1.2). It changes no state, so it takes no Idempotency-Key.
      tags: [listings]
      x-story-ids: [US-00-004, US-00-007]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleCheckCreate' }
      responses:
        '200':
          description: The rules' verdict on the draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleCheckResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: catalift_session
      description: Server-side session set by POST /v1/sessions (ADR-0006).
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      description: Date version, e.g. 2026-10-04. Missing means the server default (the newest).
      schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client UUID. Same key and body replays the first result for 24 hours; same key and a different body is 409 idempotency_conflict.
      schema: { type: string, format: uuid }
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: The csrf_token from the session; required on every state-changing request (ADR-0006).
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from page.next_cursor.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    UploadFilter:
      name: filter[upload_id]
      in: query
      description: Only this upload's records; the screens pass the latest upload (v1.7).
      schema: { type: string, pattern: '^[0-9]+$' }
...
    ListingId:
      name: listing_id
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9]+$' }
...
  schemas:
    Error:
      type: object
      required: [error]
      x-error-codes:
        - { code: validation_failed, status: 400, meaning: "Malformed request, an unknown filter, or a field failed validation; details names the field." }
        - { code: unauthorized, status: 401, meaning: "No session, an expired one, or wrong email or password at sign-in." }
        - { code: forbidden_role, status: 403, meaning: "Signed in, but this action needs the reviewer role (AC-US-00-009-5)." }
        - { code: csrf_failed, status: 403, meaning: "X-CSRF-Token missing or not this session's token." }
        - { code: not_found, status: 404, meaning: "No such record." }
        - { code: idempotency_conflict, status: 409, meaning: "Idempotency-Key reused with a different body." }
        - { code: version_conflict, status: 409, meaning: "The listing or attributes changed since the version sent; reload and retry." }
        - { code: payload_too_large, status: 413, meaning: "CSV over 5 MB or an image over 10 MB." }
        - { code: unsupported_media_type, status: 415, meaning: "Not CSV, or an image that is not JPEG, PNG or WebP." }
        - { code: unprocessable, status: 422, meaning: "Valid shape but the content breaks a rule; details names it." }
        - { code: voice_note_required, status: 422, meaning: "A brand in scope has no voice note and neutral_voice_confirmed is false (AC-US-00-003-5)." }
        - { code: budget_blocked, status: 422, meaning: "AI calls are blocked until the owner clears the budget block (D13)." }
        - { code: listing_not_generated, status: 422, meaning: "The listing has no generated text yet, so a field cannot be regenerated." }
        - { code: no_approved_listings, status: 422, meaning: "No channel has an approved listing, so no file would be written (AC-US-00-010-5)." }
        - { code: rate_limited, status: 429, meaning: "Too many requests; wait Retry-After seconds." }
        - { code: internal, status: 500, meaning: "Unexpected failure; quote request_id to support." }
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  reason: { type: string }
            request_id: { type: string }
      example:
        error:
          code: voice_note_required
          message: Brand Indigo Loom has no voice note; add one or confirm a neutral voice.
          details: [{ field: neutral_voice_confirmed, reason: must be true when a brand has no voice note }]
          request_id: req_01J9Z6
    Page:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
      example: { next_cursor: eyJza3UiOiJLVS0xMDQifQ, has_more: true }
...
    ListingStatus:
      type: string
      enum: [queued, generated, failed, stopped_budget]
      example: generated
    RuleStatus:
      type: string
      enum: [unchecked, passing, failing]
      example: passing
    RegenerationStatus:
      type: string
      enum: [queued, applied, superseded, failed, stopped_budget]
      example: superseded
...
    RuleCheckCreate:
      type: object
      required: [listing_id]
      properties:
        listing_id: { type: string }
        title: { type: string, maxLength: 500 }
        bullet_1: { type: string, maxLength: 1000 }
        bullet_2: { type: string, maxLength: 1000 }
        bullet_3: { type: string, maxLength: 1000 }
        bullet_4: { type: string, maxLength: 1000 }
        bullet_5: { type: string, maxLength: 1000 }
        description: { type: string, maxLength: 10000 }
      example: { listing_id: '412', title: "Women's Navy Cotton Kurta, Regular Fit" }
    RuleCheckResult:
      type: object
      required: [rule_status, rule_failures]
      properties:
        rule_status: { type: string, enum: [passing, failing] }
        rule_failures: { type: array, items: { $ref: '#/components/schemas/RuleFailure' } }
      example:
        rule_status: failing
        rule_failures: [{ rule: banned_word, field: bullet_1, message: 'Bullet 1 uses the banned phrase "guaranteed".' }]
...
    RuleFailure:
      type: object
      required: [rule, field, message]
      properties:
        rule: { type: string }
        field: { type: string }
        message: { type: string }
      example: { rule: title_max_length, field: title, message: 'Title is 212 characters; the limit is 200.' }
    Listing:
      type: object
      required: [id, product_id, sku, channel, status, failure_reason, version, title, bullets, description, rule_status, rule_failures, approved, approved_at, approved_by, attributes, attributes_revision, latest_regeneration]
      properties:
        id: { type: string }
        product_id: { type: string }
        sku: { type: string }
        channel: { type: string }
        status: { $ref: '#/components/schemas/ListingStatus' }
        failure_reason: { type: [string, 'null'] }
        version: { type: integer, minimum: 0 }
        title: { type: [string, 'null'] }
        bullets:
          type: [array, 'null']
          minItems: 5
          maxItems: 5
          items: { type: string }
        description: { type: [string, 'null'] }
        rule_status: { $ref: '#/components/schemas/RuleStatus' }
        rule_failures: { type: array, items: { $ref: '#/components/schemas/RuleFailure' } }
        approved: { type: boolean, description: True only while an approval matches the current version. }
        approved_at: { type: [string, 'null'], format: date-time }
        approved_by: { type: [string, 'null'], description: Email of the approving reviewer. }
        attributes:
          type: object
          description: The product's current attributes, which the reviewer checks the text against (v1.1).
          required: [colour, pattern, sleeve, neckline, fit]
          properties:
            colour: { type: [string, 'null'] }
            pattern: { type: [string, 'null'] }
            sleeve: { type: [string, 'null'] }
            neckline: { type: [string, 'null'] }
            fit: { type: [string, 'null'] }
        attributes_revision: { type: integer, minimum: 0, description: 'Send it to PATCH /products/{product_id}/attributes to correct them (v1.1).' }
        latest_regeneration:
          type: [object, 'null']
          description: The newest regeneration request on this listing, so the grid can show a pending or superseded one (v1.1).
          required: [field, status]
          properties:
            field: { type: string }
            status: { $ref: '#/components/schemas/RegenerationStatus' }
      example:
        id: '412'
        product_id: '104'
        sku: KU-104
        channel: amazon_style
        status: generated
        failure_reason: null
        version: 3
        title: Women's Navy Cotton Kurta, Regular Fit, Three-Quarter Sleeve
        bullets: [Soft breathable cotton, Regular fit, Three-quarter sleeves, Round neckline, Machine washable]
        description: A navy cotton kurta cut for everyday wear.
        rule_status: passing
        rule_failures: []
        approved: true
        approved_at: 2026-10-04T10:12:00Z
        approved_by: reviewer@catalift.example
        attributes: { colour: navy, pattern: solid, sleeve: three-quarter sleeve, neckline: round neck, fit: regular }
        attributes_revision: 2
        latest_regeneration: null
    ListingPage:
      type: object
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Listing' } }
        page: { $ref: '#/components/schemas/Page' }
      example:
        data:
          - id: '413'
            product_id: '104'
            sku: KU-104
            channel: own_website
            status: generated
            failure_reason: null
            version: 1
            title: Navy Cotton Kurta
            bullets: [Soft cotton, Regular fit, Three-quarter sleeves, Round neck, Easy care]
            description: Our everyday navy kurta.
            rule_status: failing
            rule_failures: [{ rule: banned_word, field: bullet_5, message: 'Banned word "easy care" in bullet_5.' }]
            approved: false
            approved_at: null
            approved_by: null
            attributes: { colour: navy, pattern: solid, sleeve: three-quarter sleeve, neckline: round neck, fit: regular }
            attributes_revision: 2
            latest_regeneration: { field: title, status: queued }
        page: { next_cursor: null, has_more: false }
    ListingUpdate:
      type: object
      required: [version]
      minProperties: 2
      properties:
        version: { type: integer, minimum: 1, description: The version the reviewer saw. }
        title: { type: string, minLength: 1, maxLength: 500 }
        bullet_1: { type: string, minLength: 1, maxLength: 1000 }
        bullet_2: { type: string, minLength: 1, maxLength: 1000 }
        bullet_3: { type: string, minLength: 1, maxLength: 1000 }
        bullet_4: { type: string, minLength: 1, maxLength: 1000 }
        bullet_5: { type: string, minLength: 1, maxLength: 1000 }
        description: { type: string, minLength: 1, maxLength: 10000 }
      example: { version: 2, title: 'Women''s Black Cotton Kurta, Regular Fit' }
...
    ApprovalBatchCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            type: object
            required: [listing_id, version]
            properties:
              listing_id: { type: string }
              version: { type: integer, minimum: 1 }
      example:
        items:
          - { listing_id: '412', version: 3 }
          - { listing_id: '413', version: 1 }
    ApprovalBatchResult:
      type: object
      required: [approved, skipped]
      properties:
        approved:
          type: array
          items:
            type: object
            required: [listing_id, version, approved_at]
            properties:
              listing_id: { type: string }
              version: { type: integer }
              approved_at: { type: string, format: date-time }
        skipped:
          type: array
          items:
            type: object
            required: [listing_id, reason]
            properties:
              listing_id: { type: string }
              reason: { type: string, enum: [failing_rules, version_changed, not_generated, not_found] }
      example:
        approved: [{ listing_id: '412', version: 3, approved_at: 2026-10-04T10:12:00Z }]
        skipped: [{ listing_id: '413', reason: failing_rules }]
...
  responses:
    BadRequest:
      description: Malformed request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: No session, or wrong email or password
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Wrong role or missing CSRF token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such record
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Idempotency key reused with a different body, or a stale version
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    UnprocessableEntity:
      description: Valid shape, but the content breaks a rule
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooManyRequests:
      description: Rate limited; see Retry-After
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    InternalError:
      description: Unexpected failure; request_id is in the body
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }